move tests regarding highlighting to the test modules related to the type of the subject
